% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmutation_table.R
\name{commutation_table}
\alias{commutation_table}
\title{Build an annual commutation table (discrete ages)}
\usage{
commutation_table(lt, i, check = TRUE)
}
\arguments{
\item{lt}{A lifetable object as produced by \code{\link{lifetable}}.
Must contain columns \code{x} and \code{lx}. If available, \code{qx} or \code{px}
may also be used, but \code{dx} is computed robustly from successive \code{lx}.}

\item{i}{Numeric. Annual effective interest rate (must satisfy \code{i > -1}).}

\item{check}{Logical. If \code{TRUE}, performs basic input checks.}
}
\value{
A tibble with columns \code{x}, \code{lx}, \code{dx}, \code{v},
\code{Dx}, \code{Nx}, \code{Sx}, \code{Cx}, \code{Mx}, \code{Rx}.
}
\description{
Constructs classical annual commutation functions Dx, Nx, Sx, Cx, Mx, Rx
from a lifetable defined at integer ages and an annual effective interest rate i.
}
