% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_reserve.R
\name{.mc_reserve_benefit_info}
\alias{.mc_reserve_benefit_info}
\title{Internal helper: benefit payment information for reserve calculations}
\usage{
.mc_reserve_benefit_info(Kx, Tx, death_time, type, n, h, timing)
}
\description{
Internal helper: benefit payment information for reserve calculations
}
\keyword{internal}
