% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_life_mc.R
\name{.validate_mc_common}
\alias{.validate_mc_common}
\title{Internal helper: validate common Monte Carlo arguments}
\usage{
.validate_mc_common(x, i, m, n_sim, seed, n)
}
\description{
Internal helper: validate common Monte Carlo arguments
}
\keyword{internal}
