% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-samples-roxygen.R, R/data-samples.R
\docType{data}
\name{bonds_sample}
\alias{bonds_sample}
\title{Sample bond contracts for fixed-income examples}
\format{
A tibble with 5 rows and 8 variables:
\describe{
\item{bond_id}{Bond identifier.}
\item{face_value}{Face value of the bond.}
\item{coupon_rate}{Annual coupon rate.}
\item{coupon_frequency}{Number of coupon payments per year.}
\item{maturity_years}{Maturity in years.}
\item{yield_rate}{Annual nominal yield rate consistent with coupon frequency.}
\item{bond_type}{Short bond type label.}
\item{price}{Theoretical bond price computed from the listed yield rate.}
}

A tibble with 5 rows and 8 variables:
\describe{
\item{bond_id}{Bond identifier.}
\item{face_value}{Face value of the bond.}
\item{coupon_rate}{Annual coupon rate.}
\item{coupon_frequency}{Number of coupon payments per year.}
\item{maturity_years}{Maturity in years.}
\item{yield_rate}{Annual nominal yield rate consistent with coupon frequency.}
\item{bond_type}{Short bond type label.}
\item{price}{Theoretical bond price computed from the listed yield rate.}
}
}
\source{
Synthetic pedagogical data created for tidyactuarial examples.

Synthetic pedagogical data created for tidyactuarial examples.
}
\usage{
bonds_sample

bonds_sample
}
\description{
A small pedagogical dataset containing bond contracts for pricing,
yield-to-maturity, duration, and convexity examples.

A small pedagogical dataset containing bond contracts for pricing,
yield-to-maturity, duration, and convexity examples.
}
\examples{
data(bonds_sample)

bonds_sample |>
  dplyr::select(bond_id, face_value, coupon_rate, maturity_years, price)

data(bonds_sample)

bonds_sample |>
  dplyr::select(bond_id, face_value, coupon_rate, maturity_years, price)

}
\keyword{datasets}
