% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-samples-roxygen.R, R/data-samples.R
\docType{data}
\name{cash_flows_sample}
\alias{cash_flows_sample}
\title{Sample cash flows for interest theory examples}
\format{
A tibble with 19 rows and 5 variables:
\describe{
\item{scenario_id}{Scenario identifier.}
\item{time}{Payment time.}
\item{amount}{Cash-flow amount. Negative values represent outflows and
positive values represent inflows.}
\item{cashflow_type}{Type of cash flow.}
\item{description}{Short description of the cash flow.}
}

A tibble with 19 rows and 5 variables:
\describe{
\item{scenario_id}{Scenario identifier.}
\item{time}{Payment time.}
\item{amount}{Cash-flow amount. Negative values represent outflows and
positive values represent inflows.}
\item{cashflow_type}{Type of cash flow.}
\item{description}{Short description of the cash flow.}
}
}
\source{
Synthetic pedagogical data created for tidyactuarial examples.

Synthetic pedagogical data created for tidyactuarial examples.
}
\usage{
cash_flows_sample

cash_flows_sample
}
\description{
A small pedagogical dataset containing cash-flow scenarios for present value,
future value, net present value, internal rate of return, equations of value,
and cash-flow diagrams.

A small pedagogical dataset containing cash-flow scenarios for present value,
future value, net present value, internal rate of return, equations of value,
and cash-flow diagrams.
}
\examples{
data(cash_flows_sample)

cash_flows_sample |>
  dplyr::filter(scenario_id == "investment_project")

data(cash_flows_sample)

cash_flows_sample |>
  dplyr::filter(scenario_id == "investment_project")

}
\keyword{datasets}
