% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utils.R
\name{.mc_assert_character_scalar}
\alias{.mc_assert_character_scalar}
\title{Internal helper: validate character scalar}
\usage{
.mc_assert_character_scalar(x, arg)
}
\arguments{
\item{x}{Object to validate.}

\item{arg}{Character string with the argument name.}
}
\value{
Invisibly returns \code{TRUE} if validation is successful.
}
\description{
Internal helper: validate character scalar
}
\keyword{internal}
