% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-samples-roxygen.R, R/data-samples.R
\docType{data}
\name{loans_sample}
\alias{loans_sample}
\title{Sample loan contracts for amortization examples}
\format{
A tibble with 4 rows and 7 variables:
\describe{
\item{loan_id}{Loan identifier.}
\item{principal}{Initial loan principal.}
\item{annual_effective_rate}{Annual effective interest rate.}
\item{term_months}{Loan term in months.}
\item{payments_per_year}{Number of payments per year.}
\item{loan_type}{Short loan type label.}
\item{payment}{Level payment amount.}
}

A tibble with 4 rows and 7 variables:
\describe{
\item{loan_id}{Loan identifier.}
\item{principal}{Initial loan principal.}
\item{annual_effective_rate}{Annual effective interest rate.}
\item{term_months}{Loan term in months.}
\item{payments_per_year}{Number of payments per year.}
\item{loan_type}{Short loan type label.}
\item{payment}{Level payment amount.}
}
}
\source{
Synthetic pedagogical data created for tidyactuarial examples.

Synthetic pedagogical data created for tidyactuarial examples.
}
\usage{
loans_sample

loans_sample
}
\description{
A small pedagogical dataset containing level-payment loan contracts for
amortization schedule and outstanding balance examples.

A small pedagogical dataset containing level-payment loan contracts for
amortization schedule and outstanding balance examples.
}
\examples{
data(loans_sample)

loans_sample |>
  dplyr::select(loan_id, principal, annual_effective_rate, term_months, payment)

data(loans_sample)

loans_sample |>
  dplyr::select(loan_id, principal, annual_effective_rate, term_months, payment)

}
\keyword{datasets}
